+++ changelog.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased] - 2017-11-28 Yudhvir
+### Added
+- Populate database with lower mainland and provinces of canada
+
 ## [Unreleased] - 2017-10-20 Stephen
 ### Added
 - Raycast algorithm implemented